requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
blis==1.3.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
cymem==2.0.11
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
django-extensions==4.1
django-rest-passwordreset==1.5.0
django-unfold==0.58.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-yasg==1.21.10
fr_core_news_md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.8.0/fr_core_news_md-3.8.0-py3-none-any.whl#sha256=8a70d090a54ef77525c3ffa6a6195b9d365f2cf369ae1cd84ede93f3d709079e
frozenlist==1.6.0
graphviz==0.20.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mongoengine==0.29.1
multidict==6.4.3
murmurhash==1.0.12
numpy==2.2.5
openai==0.28.0
packaging==25.0
preshed==3.0.9
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
pydot==4.0.0
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.12.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.0
setuptools==80.4.0
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
swagger-spec-validator==3.0.4
thinc==8.3.6
tqdm==4.67.1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
yarl==1.20.0
